Ensure description exists for HTTP error.
Don't fail if the HTTP error raised with json payload but
not in the expect format.

For example, the json returned by Keystone is not the same
as the Gnocchi one.

Any middleware between Gnocchi and the Webserver can potentialy
return json.

This change handles these cases gracefully instead of raising
a TypeError.

Closes-bug: #1543449

Change-Id: Id824ed4753271a4451881656293b53132ff3eb84