Ensure description exists for HTTP error.
Don't fail if the HTTP error raised with json payload but
not in the expect format.

The regular expression that parse the description expect to get a
string. But if "description" is missing, we get None and the re.match()
raise "TypeError: expected string or buffer"

This change fixes that.

Change-Id: Id824ed4753271a4451881656293b53132ff3eb84