Stop always encoding resource
This makes sure we do not try to guess the mechanism used by the server if it
ever changes.

The test is changed to not make UUID with / in it working. This won't
work for a lot of reasons, and it should probably not work, so it's the
responsibility of the caller to not use a / in the resource id.

Change-Id: Ic7ce4033ee13c65e6f542f52ebf15d8f271deee6