resource: Set correct attribute type
Currently all resources attributes are transmitted as string to the
server. But if the attribute type is a number or a bool, the json must
have the value as float or bool too.

This change does that.

Closes-bug: #1649269
Change-Id: I0f739f782bf081761d289398bcd4185af3d0e7b4
(cherry picked from commit c77109d8b3c074ddc80daa1513c6b12ce55bac00)