Fix data display when listing resources
The list of columns is currently hardcoded and does not take into account extra
columns returned when using the `--details` option. This build a dynamic column
list in all cases.

It also hides the (deprecated) fields `created_by_*` if `creator` is present.
It also hides the `metrics` field in most cases (except show) as it's usually
big and make the list very long in some cases.

Change-Id: Id25e9f30189cb7faccf3f4b9adb4df413435296c